requirements.txt
flask
requests
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-requests
